Update security for lab 3.

+++ security/11-training-php/list_users.php
@@ -1,7 +1,6 @@
 <?php
 // Start the session
 session_start();
-
 require_once 'models/UserModel.php';
 $userModel = new UserModel();
 
@@ -15,17 +14,31 @@ $users = $userModel->getUsers($params);
 <!DOCTYPE html>
 <html>
 <head>
-    <title>Home</title>
+    <title>List of Users</title>
     <?php include 'views/meta.php' ?>
 </head>
 <body>
     <?php include 'views/header.php'?>
     <div class="container">
-        <?php if (!empty($users)) {?>
-            <div class="alert alert-warning" role="alert">
-                List of users! <br>
-                Hacker: http://php.local/list_users.php?keyword=ASDF%25%22%3BTRUNCATE+banks%3B%23%23
+        <?php if (isset($_SESSION['success'])): ?>
+            <div class="alert alert-success">
+                <?php echo $_SESSION['success']; ?>
+                <?php unset($_SESSION['success']); // Gỡ thông báo khỏi session ?>
             </div>
+        <?php endif; ?>
+
+        <?php if (!empty($users)) { ?>
+            <div class="alert alert-warning" role="alert">
+                List of users!
+            </div>
+
+            <?php if (isset($_SESSION['error'])): ?>
+                <div class="alert alert-danger">
+                    <?php echo $_SESSION['error']; ?>
+                    <?php unset($_SESSION['error']); // Gỡ thông báo khỏi session ?>
+                </div>
+            <?php endif; ?>
+            
             <table class="table table-striped">
                 <thead>
                     <tr>
@@ -37,26 +50,21 @@ $users = $userModel->getUsers($params);
                     </tr>
                 </thead>
                 <tbody>
-                    <?php foreach ($users as $user) {?>
+                    <?php foreach ($users as $user) { ?>
                         <tr>
-                            <th scope="row"><?php echo $user['id']?></th>
+                            <!-- Sử dụng htmlspecialchars để ngăn chặn sử dụng script để thâm nhập đường dẫn của hacker -->
+                            <th scope="row"><?php echo htmlspecialchars($user['id']); ?></th> 
+                            <td><?php echo htmlspecialchars($user['name']); ?></td>
+                            <td><?php echo htmlspecialchars($user['fullname']); ?></td>
+                            <td><?php echo htmlspecialchars($user['type']); ?></td>
                             <td>
-                                <?php echo $user['name']?>
-                            </td>
-                            <td>
-                                <?php echo $user['fullname']?>
-                            </td>
-                            <td>
-                                <?php echo $user['type']?>
-                            </td>
-                            <td>
-                                <a href="form_user.php?id=<?php echo $user['id'] ?>">
+                                <a href="form_user.php?id=<?php echo base64_encode($user['id']); ?>">
                                     <i class="fa fa-pencil-square-o" aria-hidden="true" title="Update"></i>
                                 </a>
-                                <a href="view_user.php?id=<?php echo $user['id'] ?>">
+                                <a href="view_user.php?id=<?php echo base64_encode($user['id']); ?>">
                                     <i class="fa fa-eye" aria-hidden="true" title="View"></i>
                                 </a>
-                                <a href="delete_user.php?id=<?php echo $user['id'] ?>">
+                                <a href="delete_user.php?id=<?php echo base64_encode($user['id']); ?>" onclick="return confirm('Are you sure you want to delete this user?');">
                                     <i class="fa fa-eraser" aria-hidden="true" title="Delete"></i>
                                 </a>
                             </td>
@@ -64,9 +72,9 @@ $users = $userModel->getUsers($params);
                     <?php } ?>
                 </tbody>
             </table>
-        <?php }else { ?>
+        <?php } else { ?>
             <div class="alert alert-dark" role="alert">
-                This is a dark alert—check it out!
+                No users found.
             </div>
         <?php } ?>
     </div>
